Use VADRMPRIMESurfaceDescriptor to create surface
This change is to replace VASurfaceAttribExternalBuffers with
VADRMPRIMESurfaceDescriptor in va surface creation. With
VADRMPRIMESurfaceDescriptor, the tile info can be specified
explicitly and sent to media driver.

width iGPU, tileY is being used by default and there is no need
to set this info, but with dGPU, tile info should be set.

Tracked-On: OAM-108862
Signed-off-by: Chen, Tianmi <tianmi.chen@intel.com>