Add minimum limit to BufferSizeInKB for corner cases
The BufferSizeInKB is estimated in oneVPL if set to 0 when calling encoder Init().
However, the estimated value is not sufficient in some corner cases, such as HEVC
with a bitrate of 4 and a resolution of 176x144. In these cases, set BufferSizeInKB
manually to a limit to ensure there is enough space to write the bitstream.
cases: android.videocodec.cts.VideoEncoderMinMaxTest#testMinMaxSupport
[63_c2.intel.hevc.encoder_video/hevc_0.00mbps_176x144_30fps_maxb-0_vbr_i-dist-0]

Tracked-On: OAM-118627
Signed-off-by: Nana Zhang <nana.zhang@intel.com>