Fix detach surface cts test run timeout issue
The issue is detach surface cts test run timeout, cause is it
keeps reallocating surface when the allocated surface is not in
cached buffer pool. After removing the InCache() limitation,
sometimes we get a locked block which is still used by onevpl and
will cause random fail.

Solution is 1) Remove code to check if allocated surfce is in
cache. 2) For detached surface situation, when get locked surface
we keep fetching new surface from framework until we get an
unlocked one.

Tracked-On: OAM-131939
Signed-off-by: Lina Sun <lina.sun@intel.com>
Signed-off-by: zhangyichix <yichix.zhang@intel.com>